Add a function for ambient authorities known at compile time.
Sometimes, one is primarily concerned with resource names that may be
influenced by untrusted code at runtime, and compile-time names and
addresses are ok.

Add a new function for obtaining an `AmbientAuthority`, meant for use
by macros which confirm that their operands are literals known at
compile time. Add it to the clippy config so that we still warn about
it by default, but this way, users have the option of commenting it
out if they wish to focus on dynamic ambient authority.

This implements the ambient-authority side of bytecodealliance/cap-std#175.